+++ src/components/content/tab-item/tab-item.vue
@@ -15,6 +15,16 @@ import { nanoid } from "nanoid";
 
 const props = defineProps({
 	/**
+	 * Any ID to apply to this tab. This can be used in conjunction with
+	 * `tab-group` prop `updateUrl` to reinstate tabs on page refresh. When
+	 * providing an ID, ensure that it is unique.
+	 */
+	id: {
+		type: String,
+		default: null,
+	},
+
+	/**
 	 * Whether this tab is initially active.
 	 */
 	initiallyActive: {
@@ -40,7 +50,7 @@ const active = computed(() => activeTabId.value === tabId.value);
 // `component`.
 const label = computed(() => slots.label);
 // The ID of this tab, used to link the panel and the tab together.
-const tabId = computed(() => nanoid());
+const tabId = computed(() => (isNonEmptyString(props.id) ? props.id : `tab-${nanoid()}`));
 // The ID of this tab, used to link the panel and the tab together.
 const panelId = computed(() => `${tabId.value}-panel`);
 
@@ -54,7 +64,8 @@ onMounted(() => {
 		icon: props.icon,
 	}, active);
 
-	// If there is a hash
+	// If there is a hash in the URL, check if it matches a tab, or the contents
+	// of this tab. If so, activate this tab.
 	const hash = window.location.hash.slice(1);
 
 	if (!isNonEmptyString(hash)) {