+++ src/components/interaction/copy-content/copy-content.test.js
@@ -2,7 +2,8 @@ import { createMount } from "@unit/support/mount";
 import { describe, expect, test } from "vitest";
 import CopyContent from "./copy-content.vue";
 
-const mount = createMount(CopyContent);
+const defaultProps = { content: "Content to copy" };
+const mount = createMount(CopyContent, { props: defaultProps });
 
 describe("copy-content", () => {
 	describe("Initialisation", () => {