+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This version introduces a new documentation site to demonstrate each component a
 
 - `tab-group`: Icons can now be added to tabs via the `icon` prop on a `tab-item`.
 - `tab-group`: When the URL contains a fragment, if that fragment is contained within a tab, it will activate itself.
+- `tab-group`: A new `updateUrl` prop allows an active tab to be reinstated on page reload.
 
 # 0.22.0 - 2025-04-18